requirements.txt
fastapi==0.115.6
httpx==0.28.1
motor==3.6.0
piexif==1.1.3
Pillow==11.0.0
uvicorn==0.34.0
